refactor: make changes to integrate into hedwig
