+++ package.json
@@ -16,6 +16,7 @@
     "build-min": "webpack --config ./config/webpack.config.js --env=min",
     "build": "webpack --config ./config/webpack.config.js --env=dev",
     "build-module": "webpack --config ./config/webpack.config.js --env=dev-module && rimraf ./tmp && tsc --outDir ./tmp && babel tmp --quiet --out-dir dist/module && copyfiles -u 2 dist/dev-module/*.css dist/module && copyfiles -u 2 dist/dev-module/*.bundle.* dist/module && rimraf ./tmp",
+    "build-module-copy-to-ng-hedwig": "npm run build-module && cp dist/module/{*.bundle.js,main.css} ../ng-hedwig/",
     "build-python": "webpack --config ./config/webpack.config.js --env=python-dev",
     "build-python-min": "webpack --config ./config/webpack.config.js --env=python-min",
     "build:watch": "webpack --config ./config/webpack.config.js --watch",
