+++ src/neuroglancer/segmentation_user_layer.ts
@@ -203,10 +203,9 @@ export class SegmentationUserLayer extends Base {
       if (this.addStaticAnnotations(loadedSubsource)) continue;
       const {volume, mesh, segmentPropertyMap} = loadedSubsource.subsourceEntry.subsource;
       if (volume instanceof MultiscaleVolumeChunkSource) {
-        switch (volume.dataType) {
-          case DataType.FLOAT32:
-            loadedSubsource.deactivate('Data type not compatible with segmentation layer');
-            continue;
+        if (volume.dataType === DataType.FLOAT32) {
+          loadedSubsource.deactivate('Data type not compatible with segmentation layer');
+          continue;
         }
         loadedSubsource.activate(
             () => loadedSubsource.addRenderLayer(new SegmentationRenderLayer(volume, {
